+++ src/main/java/com/saintdan/framework/controller/management/ClientController.java
@@ -3,7 +3,6 @@ package com.saintdan.framework.controller.management;
 import com.saintdan.framework.annotation.CurrentUser;
 import com.saintdan.framework.component.ResultHelper;
 import com.saintdan.framework.component.ValidateHelper;
-import com.saintdan.framework.constant.CommonsConstant;
 import com.saintdan.framework.constant.ResourceURL;
 import com.saintdan.framework.constant.VersionConstant;
 import com.saintdan.framework.domain.ClientDomain;
@@ -11,23 +10,15 @@ import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationType;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.ClientParam;
-import com.saintdan.framework.po.Client;
 import com.saintdan.framework.po.User;
-import com.saintdan.framework.tools.QueryHelper;
 import com.saintdan.framework.vo.ClientVO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
-import net.kaczmarzyk.spring.data.jpa.domain.In;
-import net.kaczmarzyk.spring.data.jpa.domain.Like;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -81,49 +72,14 @@ public class ClientController {
   /**
    * Show all.
    *
-   * @param param {@link ClientParam}
    * @return all clients.
    */
   @RequestMapping(method = RequestMethod.GET)
   @ApiOperation(value = "List", httpMethod = "GET", response = ClientVO.class)
-  @ApiImplicitParams({
-      @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true),
-      @ApiImplicitParam(name = "name", value = "client's name", paramType = "query", dataType = "string"),
-      @ApiImplicitParam(name = "pageNo", dataType = "date", paramType = "query"),
-      @ApiImplicitParam(name = "pageSize", dataType = "date", paramType = "query"),
-      @ApiImplicitParam(name = "sortBy", dataType = "date", paramType = "query", example = "sortBy=id:desc,username:desc")
-  })
-  public ResponseEntity all(
-      @And({
-          @Spec(path = "name", spec = Like.class),
-          @Spec(path = "validFlag", constVal = "VALID", spec = In.class)
-      }) @ApiIgnore Specification<Client> clientSpecification, @ApiIgnore ClientParam param) {
+  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
+  public ResponseEntity all() {
     try {
-      if (param.getPageNo() == null) {
-        return new ResponseEntity<>(clientDomain.getAll(clientSpecification, QueryHelper.getSort(param.getSortBy()), ClientVO.class), HttpStatus.OK);
-      }
-      return new ResponseEntity<>(clientDomain.getPage(clientSpecification, QueryHelper.getPageRequest(param), ClientVO.class), HttpStatus.OK);
-    } catch (Exception e) {
-      // Return unknown error and log the exception.
-      return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-  }
-
-  /**
-   * Show {@link com.saintdan.framework.vo.ClientVO} by id.
-   *
-   * @param id id of client
-   * @return {@link com.saintdan.framework.vo.ClientVO}
-   */
-  @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-  @ApiOperation(value = "Detail", httpMethod = "GET", response = ClientVO.class)
-  @ApiImplicitParam(name = "id", value = "client's id", paramType = "path", dataType = "string", required = true)
-  public ResponseEntity detail(@ApiIgnore @PathVariable String id) {
-    try {
-      if (StringUtils.isBlank(id)) {
-        return resultHelper.infoResp(ErrorType.SYS0002, CommonsConstant.ID_BLANK, HttpStatus.UNPROCESSABLE_ENTITY);
-      }
-      return new ResponseEntity<>(clientDomain.getById(Long.valueOf(id), ResponseEntity.class), HttpStatus.OK);
+      return new ResponseEntity<>(clientDomain.all(), HttpStatus.OK);
     } catch (Exception e) {
       // Return unknown error and log the exception.
       return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
@@ -136,8 +92,12 @@ public class ClientController {
    * @param param {@link ClientParam}
    * @return {@link com.saintdan.framework.vo.ClientVO}
    */
-  @RequestMapping(value = "/{id}", method = RequestMethod.POST)
+  @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
   @ApiOperation(value = "Update", httpMethod = "PUT", response = ClientVO.class)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "id", value = "client's id", paramType = "path", dataType = "long", required = true)
+  })
   public ResponseEntity update(@ApiIgnore @CurrentUser User currentUser, @RequestBody ClientParam param) {
     try {
       // Validate current user, param and sign.
@@ -146,7 +106,7 @@ public class ClientController {
         return responseEntity;
       }
       // Update client.
-      return new ResponseEntity<>(clientDomain.update(ClientVO.class, param, currentUser), HttpStatus.OK);
+      return new ResponseEntity<>(clientDomain.update(param, currentUser), HttpStatus.OK);
     } catch (CommonsException e) {
       // Return error information and log the exception.
       return resultHelper.infoResp(logger, e.getErrorType(), e.getMessage(), HttpStatus.UNPROCESSABLE_ENTITY);
@@ -159,21 +119,25 @@ public class ClientController {
   /**
    * Delete client.
    *
-   * @param param {@link ClientParam}
+   * @param id {@link com.saintdan.framework.po.Client#id}
    * @return {@link com.saintdan.framework.vo.ClientVO}
    */
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
   @ApiOperation(value = "Delete", httpMethod = "DELETE", response = ResponseEntity.class)
-  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
-  public ResponseEntity delete(@ApiIgnore @CurrentUser User currentUser, @RequestBody ClientParam param) {
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "id", value = "client's id", paramType = "path", dataType = "long", required = true)
+  })
+  public ResponseEntity delete(@ApiIgnore @CurrentUser User currentUser, @ApiIgnore @PathVariable Long id) {
     try {
+      ClientParam param = new ClientParam(id);
       // Validate current user and param.
       ResponseEntity responseEntity = validateHelper.validate(param, currentUser, logger, OperationType.DELETE);
       if (!responseEntity.getStatusCode().is2xxSuccessful()) {
         return responseEntity;
       }
       // Delete client.
-      clientDomain.delete(param, currentUser);
+      clientDomain.deepDelete(param.getId(), currentUser);
       return new ResponseEntity(HttpStatus.NO_CONTENT);
     } catch (CommonsException e) {
       // Return error information and log the exception.