+++ src/main/java/com/saintdan/framework/controller/management/ResourceController.java
@@ -11,23 +11,16 @@ import com.saintdan.framework.enums.ErrorType;
 import com.saintdan.framework.enums.OperationType;
 import com.saintdan.framework.exception.CommonsException;
 import com.saintdan.framework.param.ResourceParam;
-import com.saintdan.framework.po.Resource;
 import com.saintdan.framework.po.User;
-import com.saintdan.framework.tools.QueryHelper;
 import com.saintdan.framework.vo.ResourceVO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
-import net.kaczmarzyk.spring.data.jpa.domain.In;
-import net.kaczmarzyk.spring.data.jpa.domain.Like;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.And;
-import net.kaczmarzyk.spring.data.jpa.web.annotation.Spec;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -73,24 +66,10 @@ import springfox.documentation.annotations.ApiIgnore;
 
   @RequestMapping(method = RequestMethod.GET)
   @ApiOperation(value = "List", httpMethod = "GET", response = ResourceVO.class)
-  @ApiImplicitParams({
-      @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true),
-      @ApiImplicitParam(name = "name", value = "resource's name", paramType = "query", dataType = "string"),
-      @ApiImplicitParam(name = "pageNo", dataType = "date", paramType = "query"),
-      @ApiImplicitParam(name = "pageSize", dataType = "date", paramType = "query"),
-      @ApiImplicitParam(name = "sortBy", dataType = "date", paramType = "query", example = "sortBy=id:desc,username:desc")
-  })
-  public ResponseEntity all(
-      @And({
-          @Spec(path = "name", spec = Like.class),
-          @Spec(path = "path", spec = Like.class),
-          @Spec(path = "validFlag", constVal = "VALID", spec = In.class)
-      }) @ApiIgnore Specification<Resource> resourceSpecification, @ApiIgnore ResourceParam param) {
+  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
+  public ResponseEntity all() {
     try {
-      if (param.getPageNo() == null) {
-        return new ResponseEntity<>(resourceDomain.getAll(resourceSpecification, QueryHelper.getSort(param.getSortBy()), ResourceVO.class), HttpStatus.OK);
-      }
-      return new ResponseEntity<>(resourceDomain.getPage(resourceSpecification, QueryHelper.getPageRequest(param), ResourceVO.class), HttpStatus.OK);
+      return new ResponseEntity<>(resourceDomain.all(), HttpStatus.OK);
     } catch (Exception e) {
       // Return unknown error and log the exception.
       return resultHelper.errorResp(logger, e, ErrorType.UNKNOWN, e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
@@ -99,7 +78,10 @@ import springfox.documentation.annotations.ApiIgnore;
 
   @RequestMapping(value = "/{id}", method = RequestMethod.GET)
   @ApiOperation(value = "Detail", httpMethod = "GET", response = ResourceVO.class)
-  @ApiImplicitParam(name = "id", value = "resource's id", paramType = "path", dataType = "string", required = true)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "id", paramType = "path", dataType = "string", required = true)
+  })
   public ResponseEntity detail(@ApiIgnore @PathVariable String id) {
     try {
       if (StringUtils.isBlank(id)) {
@@ -114,6 +96,10 @@ import springfox.documentation.annotations.ApiIgnore;
 
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
   @ApiOperation(value = "Update", httpMethod = "PUT", response = ResourceVO.class)
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "id", paramType = "path", dataType = "string", required = true)
+  })
   public ResponseEntity update(@ApiIgnore @CurrentUser User currentUser, @RequestBody ResourceParam param) {
     try {
       // Validate current user, param and sign.
@@ -134,16 +120,20 @@ import springfox.documentation.annotations.ApiIgnore;
 
   @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
   @ApiOperation(value = "Delete", httpMethod = "DELETE", response = ResponseEntity.class)
-  @ApiImplicitParam(name = "Authorization", value = "token", paramType = "header", dataType = "string", required = true)
-  public ResponseEntity delete(@CurrentUser User currentUser, @RequestBody ResourceParam param) {
+  @ApiImplicitParams({
+      @ApiImplicitParam(name = "Authorization", paramType = "header", dataType = "string", required = true),
+      @ApiImplicitParam(name = "id", paramType = "path", dataType = "string", required = true)
+  })
+  public ResponseEntity delete(@ApiIgnore @CurrentUser User currentUser, @ApiIgnore @PathVariable Long id) {
     try {
+      ResourceParam param = new ResourceParam(id);
       // Validate current user and param.
       ResponseEntity responseEntity = validateHelper.validate(param, currentUser, logger, OperationType.DELETE);
       if (!responseEntity.getStatusCode().is2xxSuccessful()) {
         return responseEntity;
       }
       // Delete resource.
-      resourceDomain.delete(param, currentUser);
+      resourceDomain.deepDelete(param.getId(), currentUser);
       return new ResponseEntity(HttpStatus.NO_CONTENT);
     } catch (CommonsException e) {
       // Return error information and log the exception.